add desktop entry for personal firefox profile
the -no-remote option is required to open firefox in two separate
profiles, but it prevents a second window from being remotely connected
to the first instance, so the profiles have to be kept separate.

there then is the problem of what to do in the "default" case, I have
aliased "firefox" to the developer edition in userspace. This isn't
perfect but should be fine for now.